Move annotations to utils/
