+++ src/main/java/com/github/mkram17/bazaarutils/misc/MajorUpdateActions.java
@@ -2,7 +2,7 @@ package com.github.mkram17.bazaarutils.misc;
 
 import com.github.mkram17.bazaarutils.BazaarUtils;
 import com.github.mkram17.bazaarutils.features.gui.buttons.Bookmarks;
-import com.github.mkram17.bazaarutils.misc.autoregistration.RunOnInit;
+import com.github.mkram17.bazaarutils.utils.annotations.autoregistration.RunOnInit;
 
 public class MajorUpdateActions {
